Initiate first world
Initiate the Gazebo world with default model,
including a model for the building and the robot
associated. The provided world has some tables
in it, as well as the start script "hello.cpp"
print a default message when the world is launched.

This is part of Build My World project for Udacity
Robotics Software Engineer Nanodegree course.